Update patch set 1
Patch Set 1: Code-Review+1

The covers the only avc that I am getting on mako. I am able to "setenforce 1" with a fresh install of the rom and the latest gapps.

There may be a better way of doing this, but this was quick.

Patch-set: 1
Reviewer: Tim Clark <5259@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1