Update patch set 1
Patch Set 1:

This works, but it feels so wrong.

We really only want to give appdomain access to the active theme, but I'm not sure how to do that. 

A better solution would be to copy the theme assets out of the app at install time into /data. This would also solve the "vold kills systemserver" issue if it still exists.

Patch-set: 1