Update patch set 1
Patch Set 1:

Note: This policy domain is based on what is required by the AOSP recovery.  Obviously CM may have different requirements for its recovery.  It uses unconfined_domain() so it will inherit all of the unconfined.te permissions, but we have been removing specific permissions from unconfined in AOSP master, which is why we had to introduce a separate domain for recovery in order to keep recovery working.

Patch-set: 1
Label: Code-Review=0