Update patch set 4
Patch Set 4:

While su and root are there purposefully -- and in a way that lets users
actually know when and how they are opting to use those -- the proposed
change on mediaserver's permissions is "silent" -- and enacting it is being
made by us (with the patch), not by the user. 

People disabling SELinux... well, it sucks, but they are *opting* to disable
SELinux (and, well, they could, instead, do this change on their device
instead). If we integrate this change, we're doing it for every userdebug
build, either the user wants it or not.

Patch-set: 4