Update patch set 2
Abandoned

This change was totally misguided. .coldboot is correctly residing in the directory passed to gatekeeperd as an argument. Device-specific code in a client of gatekeeperd are the cause of it trying to read a file from /firmware/image. I'll sort this out in a device-specific policy, not here.

I think I must have been mislead by a bad printk somewhere along the line.

Patch-set: 2
Status: abandoned