Update patch set 1
Patch Set 1:

Yeah, we need to do this, but we need a solution for enabling adb over wireless.

There are only 3 shell props:

- log.* -> I'm ok with settings being able to set this
- service.adb.root
- service.adb.tcp.port

I wouldn't want system server to be able to set the second one because, on userdebug builds that would mean that escalation to system would instantly grant escalation to root.

Patch-set: 1