Update patch set 2
Patch Set 2:

When I factory reset, the lockscreen wallpaper is a link to the homescreen wallpaper, so they both change when the wallpaper is set. The Theme chooser then removes the link and creates a file with the same name when it sets a new lockscreen wallpaper. The chooser would then need to recreate that link to the homescreen wallpaper to go back to the default behavior and selinux would detect a change in the filesystem and deny it.

Patch-set: 2