Update patch set 2
Patch Set 2:

Actually, maybe you are right. I have not factory reset for a while and I thought that keyguard_wallpaper was a link to the home screen wallpaper by default. In that case, deleting keyguard_wallpaper should go back to default behavior if it doesn't exist by default. It just seems like the solution for the theme chooser would be to create a link.

Patch-set: 2